+++ personread/src/test/java/software/daveturner/personread/CucumberRunner.java
@@ -1,0 +1,13 @@
+package software.daveturner.personread;
+
+import io.cucumber.junit.Cucumber;
+import io.cucumber.junit.CucumberOptions;
+import org.junit.runner.RunWith;
+
+@RunWith(Cucumber.class)
+@CucumberOptions(
+        features = {"src/test/resources/features"},
+        plugin = {"pretty"},
+        glue = {"software.daveturner.personread.cucumber"})
+public class CucumberRunner {
+}