+++ CHANGELOG.md
@@ -4,8 +4,9 @@ All notable changes to this project will be documented in this file.
 
 
 ## [0.0.3] - 2024-08-10
-- Added PersonRead App, pointed towards read DB
- 
+- Added PersonRead App
+	Cucumber test, read API, Mongo DB
+
 ## [0.0.2] - 2024-06-26
 - Kafka Support in PersonWrite
 - Mocking Kafka publish in PersonWrite tests and Cucumber